+++ interfaces.go
@@ -1,6 +1,0 @@
-package emperror
-
-// ErrorCollection holds a list of errors.
-type ErrorCollection interface {
-	Errors() []error
-}